+++ src/io/file.hpp
@@ -3,7 +3,7 @@
 
 #include "../pch/common.hpp"
 #include "../util/toolbox.hpp"
-#include "../http/response_stream.hpp"
+#include "../http/body_stream.hpp"
 
 #define MIME_UNSET ""
 
@@ -11,7 +11,7 @@ class File {
     public:
         File(const std::string&);
 
-        int loadToBuffer(std::unique_ptr<http::IResponseStream>&);
+        int loadToBuffer(std::unique_ptr<http::IBodyStream>&);
         std::string getLastModifiedGMT() const;
         bool exists = false;
         bool isLinked = false; // True if symlink or hardlink