+++ src/io/file_tools.cpp
@@ -1,5 +1,8 @@
 #include "file_tools.hpp"
 
+#define TMP_FILE_MAX_RETRIES 50
+#define TMP_FILE_NAME_LEN 12
+
 bool doesFileExist(const std::string& path, const bool forceInDocumentRoot) {
     const bool isFile = std::filesystem::is_regular_file(path);
     std::string docRootStr = conf::DOCUMENT_ROOT.string();
@@ -130,3 +133,47 @@ void createLogDirectoryIfMissing(const std::filesystem::path& path) {
         return;
     }
 }
+
+// Generates a random string (helper for createTempFile)
+static const char randomCharset[] = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
+void genRandomString(std::string& result, size_t length) {
+    // Create randomizer
+    std::random_device rd;
+    std::mt19937 gen(rd());
+    std::uniform_int_distribution<> dist(0, sizeof(randomCharset) - 2); // Exclude null byte
+
+    // Place random characters
+    for (size_t i = 0; i < length; ++i)
+        result += randomCharset[dist(gen)];
+}
+
+// Creates and returns the full, absolute path to a new tmp file, returning true if successful
+bool createTempFile(std::string& outPath) {
+    std::string buffer;
+    buffer.reserve(TMP_FILE_NAME_LEN);
+
+    int i = 0;
+    do {
+        // Determine temp file name
+        genRandomString(buffer, TMP_FILE_NAME_LEN);
+        outPath = (conf::TMP_PATH / buffer).string();
+    } while (++i < TMP_FILE_MAX_RETRIES && std::filesystem::exists(outPath));
+
+    // Return true if the filename is available
+    std::cout << outPath << std::endl;
+    return !std::filesystem::exists(outPath);
+}
+
+// Attempts to remove the temp file, returns true if successful
+bool removeTempFile(const std::string& tmpPath) {
+    try { // Attempt to remove
+        std::filesystem::remove(tmpPath);
+    } catch (...) {
+        ERROR_LOG << "Failed to remove temp file: " << tmpPath << std::endl;
+        return false;
+    }
+    return true;
+}
+
+#undef TMP_FILE_MAX_RETRIES
+#undef TMP_FILE_NAME_LEN